OpenCV/requirements.txt
alembic==1.11.2
backports.lzma==0.0.14
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
cmaes==0.10.0
coloredlogs==15.0.1
colorlog==6.7.0
cycler==0.11.0
fonttools==4.38.0
future==0.18.3
greenlet==2.0.2
humanfriendly==10.0
hyperopt==0.2.7
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==3.4.2.17
optuna==3.3.0
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
py4j==0.10.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.19
threadpoolctl==3.1.0
tifffile==2021.11.2
tk==0.1.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4
xgboost==1.6.2
zipp==3.15.0
